requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
SQLAlchemy==2.0.31
python-dotenv==1.0.1